+++ CHANGELOG.md
@@ -101,6 +101,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `GoogleStructureConfig` for providing Structures with Google Prompt and Embedding Driver configuration.
 - Support for `claude-3-opus`, `claude-3-sonnet`, and `claude-3-haiku` in `AnthropicPromptDriver`.
 - Support for `anthropic.claude-3-sonnet-20240229-v1:0` and `anthropic.claude-3-haiku-20240307-v1:0` in `BedrockClaudePromptModelDriver`.
+- Support for `mistral.mistral-7b-instruct-v0:2` and `mistral.mixtral-8x7b-instruct-v0:1` through `BedrockMistralPromptModelDriver`.
 - `top_k` and `top_p` parameters in `AnthropicPromptDriver`.
 - Added `AnthropicImageQueryDriver` for Claude-3 multi-modal models
 - Added `AmazonBedrockImageQueryDriver` along with `BedrockClaudeImageQueryDriverModel` for Claude-3 in Bedrock support